Implement segmented fetch in the main list to browse large resultsets.
Previously the user would need to adjust or set the date bounds in the
filtering query, or call "fetch more" which is not easily discoverable
and goes only in one direction.
The "fetch previous/next segment" method is hopefully much more
intuitive and discoverable.
Also the code supporting segmented fetch has been cleaned up and
is now a better foundation for more developments in this area.