Decouple the sort order of the main listview from the filter.
Only the "Newest first/Oldest first" setting in the preferences
affects the order of the filter.
Previously fetch_segment() was taking the listview order in
consideration, but it turns out to be unreliable, because
the fact that there are next/prev segments is tied
to the sort order of the filter.