Replace some QT_TR_NOOP by QT_TRANSLATE_NOOP with an explicit context.
The translatable strings in html_editor::m_action_definitions[]
and action_line::m_descriptions[] are part of Q_OBJECT classes,
so it seems that QT_TR_NOOP should be able to find the context,
as it did, in some previous version of lupdate.
Anyway, now it doesn't, and lupdate eliminates these strings
from the TS files, leaving them untranslatable.
The workaround implemented in this commit is to use QT_TRANSLATE_NOOP
and QCoreApplication::translate() as instructed
in https://doc.qt.io/qt-5/i18n-source-translation.html
instead of QT_TR_NOOP and QObject::tr().